Make hugo-mode mappings terminal compatible
Well, this was a revelation. Return and tab were not working in
terminal Emacs and I couldn't figure out why. Emacs reported that the
enter key was "<return>" in the GUI, but "RET" in the terminal. Why is
that? It is explained here:
http://ergoemacs.org/emacs/emacs_key_notation_return_vs_RET.html

TL;DR: they're different keycodes and it has something to do with how
terminals are antiquated and stupid.

Someone should really fix that.