Store status data in buffer-local variables
The messages and documentation indicated that the status buffer would
hold a "state" and get buried when dismissed, but in reality we
rebuilt the buffer every time you called status.

This patch saves the data in a buffer-local variable and buries the
buffer when closed, making subsequent calls to status very fast, and
allowing a forced refresh with "g".