Display titles for items in the status list
To accommodate page bundles where the filename of the content item is
"index.md" or "_index.md", we now display the title of the item from
its metadata in the status list.

We were previously opening items by re-assembling an assumed path from
the item's type and its filename, which is no longer possible when the
displayed text is the title, so instead we store the full path in a
text property on the line.

The text property approach will give us a lot more flexibility in the
future since there is no real limit to how much of the original list
metadata we can store in properties.